Move client conn handler to separate file
Users of this package may wish to build their own clients that connect
to a tcp-over-websockets server. This change makes the TCP tunnelling
handler available to package users, while still keeping the
module.exports conventions used here.

- Create `tunnel`, a new module
- Move handler that tunnels an incoming TCP connection from `client`
  to `tunnel`, naming the function `tunnelTo`
- Add `tunnel.js`to manifest of files to be exported by package.json
